refactor: merging few common functions into utils folder and using them across the app
