+++ mercata/ui/src/components/WithdrawCollateral.tsx
@@ -11,7 +11,7 @@ import { Button } from "@/components/ui/button";
 import { formatUnits, parseUnits } from "ethers";
 import { WITHDRAW_COLLATERAL_FEE } from "@/lib/contants";
 import { CollateralData, NewLoanData } from "@/interface";
-import { safeParseUnits, safeParseFloat } from "@/utils/numberUtils";
+import { safeParseUnits, safeParseFloat, getHealthFactorColor, addCommasToInput } from "@/utils/numberUtils";
 
 interface WithdrawModalProps {
   withdrawLoading: boolean;
@@ -22,25 +22,6 @@ interface WithdrawModalProps {
   onWithdraw: (amount: string) => void;
   usdstBalance?: string;
 }
-
-const addCommasToInput = (value: string) => {
-  if (!value) return '';
-  const parts = value.split('.');
-  const integerPart = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-
-  if (parts.length === 2) {
-    return integerPart + '.' + parts[1];
-  }
-  return integerPart;
-};
-
-// Calculate health factor color based on value
-const getHealthFactorColor = (healthFactor: number) => {
-  if (healthFactor >= 1.5) return "text-green-600";
-  if (healthFactor >= 1.2) return "text-yellow-600";
-  if (healthFactor >= 1.0) return "text-orange-600";
-  return "text-red-600";
-};
 
 // Calculate health impact of withdrawal
 const calculateHealthImpact = (
