+++ mercata/ui/src/pages/AssetDetail.tsx
@@ -19,28 +19,11 @@ import {
   CartesianGrid,
 } from "recharts";
 import CopyButton from '@/components/ui/copy';
+import { generatePriceData } from '@/utils/numberUtils';
 
 type PricePoint = {
   date: string;
   price: string; // since you're using `formatUnits`, it's a string
-};
-
-const generatePriceData = (basePrice: number, days: number = 30) => {
-  const data = [];
-  let currentPrice = basePrice;
-
-  for (let i = 0; i < days; i++) {
-    // Random price fluctuation between -2% and +2%
-    const change = currentPrice * (Math.random() * 0.04 - 0.02);
-    currentPrice += change;
-
-    data.push({
-      date: new Date(Date.now() - (days - i) * 24 * 60 * 60 * 1000).toLocaleDateString(),
-      price: formatUnits(currentPrice?.toLocaleString("fullwide", { useGrouping: false }), 18),
-    });
-  }
-
-  return data;
 };
 
 const AssetDetail = () => {
